refactor: 전반적인 도메인 뷰 의존 분리
