+++ src/main/java/lotto/view/OutputView.java
@@ -1,8 +1,10 @@
 package lotto.view;
 
+import java.util.stream.Collectors;
+import lotto.domain.Lotto;
 import lotto.domain.LottoGroup;
 import lotto.domain.LottoRank;
-import lotto.util.LottoResult;
+import lotto.domain.LottoResult;
 
 public class OutputView {
 
@@ -10,6 +12,10 @@ public class OutputView {
   private static final String RESULT_MESSAGE = "당첨 통계";
   private static final String BOUNDARY = "---------";
   private static final String PROFIT_FORM = "총 수익률은 %.1f%%입니다.\n";
+  private static final String LOTTO_JOIN_DELIMITER = ", ";
+  private static final String LOTTO_RANK_MESSAGE_FORM = "%d개 일치%s(%d원) - %d개";
+  private static final String BONUS_MATCH_MESSAGE = ", 보너스 볼 일치";
+  private static final String BLANK = "";
 
   public static void printMessage(final String message) {
     System.out.println(message);
@@ -17,7 +23,7 @@ public class OutputView {
 
   public static void printBoughtLotto(final LottoGroup lottoGroup) {
     lottoGroup.lottoGroup()
-        .forEach(lotto -> System.out.printf(LOTTO_FORM, lotto));
+        .forEach(lotto -> System.out.printf(LOTTO_FORM, lottoNumberMessage(lotto)));
   }
 
   public static void printLottoResult(final LottoResult lottoResult) {
@@ -28,7 +34,22 @@ public class OutputView {
         .entrySet()
         .stream()
         .filter(entrySet -> entrySet.getKey() != LottoRank.NONE)
-        .forEach(entrySet -> System.out.println(entrySet.getKey().message(entrySet.getValue())));
+        .forEach(entrySet -> System.out.println(
+            lottoRankMessage(entrySet.getKey(), entrySet.getValue()))
+        );
     System.out.printf(PROFIT_FORM, lottoResult.winningProfit());
+  }
+
+  private static String lottoNumberMessage(Lotto lotto) {
+    return lotto.lottoNumbers().stream()
+        .map(lottoNumber -> String.valueOf(lottoNumber.get()))
+        .collect(Collectors.joining(LOTTO_JOIN_DELIMITER));
+  }
+
+  private static String lottoRankMessage(LottoRank lottoRank, int count) {
+    if (lottoRank == LottoRank.SECOND) {
+      return String.format(LOTTO_RANK_MESSAGE_FORM, lottoRank.matchCount(), BONUS_MATCH_MESSAGE, lottoRank.winningMoney(), count);
+    }
+    return String.format(LOTTO_RANK_MESSAGE_FORM, lottoRank.matchCount(), BLANK, lottoRank.winningMoney(), count);
   }
 }
